+++ shiny/www/shared/prism-code-editor/languages/json.js
@@ -1,0 +1,4 @@
+import { l as languageMap } from "../index-CKRNGLIi.js";
+import { b as bracketIndenting } from "../index-BZh6eCR8.js";
+languageMap.json = languageMap.json5 = languageMap.jsonp = bracketIndenting();
+//# sourceMappingURL=json.js.map