+++ shiny/www/shared/prism-code-editor/languages/yaml.js
@@ -1,0 +1,4 @@
+import { l as languageMap } from "../index-CKRNGLIi.js";
+import { b as bracketIndenting } from "../index-BZh6eCR8.js";
+languageMap.yml = languageMap.yaml = bracketIndenting({ line: "#" });
+//# sourceMappingURL=yaml.js.map